requirements.txt
absl-py                0.10.0
astunparse             1.6.3
attrs                  19.3.0
Automat                0.8.0
blinker                1.4
cachetools             4.1.1
certifi                2020.6.20
chardet                3.0.4
Click                  7.0
cloud-init             20.2
colorama               0.4.3
command-not-found      0.3
configobj              5.0.6
constantly             15.1.0
cryptography           2.8
ctypes-callable        1.0.0.post2
cycler                 0.10.0
dbus-python            1.2.16
decorator              4.4.2
distro                 1.4.0
distro-info            0.23ubuntu1
entrypoints            0.3
gast                   0.3.3
google-auth            1.21.1
google-auth-oauthlib   0.4.1
google-pasta           0.2.0
grpcio                 1.32.0
h5py                   2.10.0
httplib2               0.14.0
hyperlink              19.0.0
idna                   2.8
imageio                2.9.0
importlib-metadata     1.5.0
imutils                0.5.3
incremental            16.10.1
Jinja2                 2.10.1
joblib                 0.16.0
jsonpatch              1.22
jsonpointer            2.0
jsonschema             3.2.0
Keras                  2.4.3
Keras-Preprocessing    1.1.2
keyring                18.0.1
kiwisolver             1.2.0
language-selector      0.1
launchpadlib           1.10.13
lazr.restfulclient     0.14.2
lazr.uri               1.0.3
libtorch               1.2.0.1
Markdown               3.2.2
MarkupSafe             1.1.0
matplotlib             3.3.1
more-itertools         4.2.0
netifaces              0.10.4
networkx               2.5
numpy                  1.19.1
oauthlib               3.1.0
opencv-python          4.4.0.42
opt-einsum             3.3.0
pandas                 1.1.2
Pillow                 7.2.0
pip                    20.0.2
protobuf               3.13.0
pyasn1                 0.4.2
pyasn1-modules         0.2.1
PyGObject              3.36.0
PyHamcrest             1.9.0
PyJWT                  1.7.1
pymacaroons            0.13.0
PyNaCl                 1.3.0
pyOpenSSL              19.0.0
pyparsing              2.4.7
pyrsistent             0.15.5
pyserial               3.4
python-apt             2.0.0+ubuntu0.20.4.1
python-dateutil        2.8.1
python-debian          0.1.36ubuntu1
pytz                   2020.1
PyWavelets             1.1.1
PyYAML                 5.3.1
requests               2.22.0
requests-oauthlib      1.3.0
requests-unixsocket    0.2.0
rsa                    4.6
scikit-image           0.17.2
scikit-learn           0.23.2
scipy                  1.4.1
SecretStorage          2.3.1
service-identity       18.1.0
setuptools             45.2.0
simplejson             3.16.0
six                    1.14.0
sklearn                0.0
ssh-import-id          5.10
systemd-python         234
tensorboard            2.2.2
tensorboard-plugin-wit 1.7.0
tensorflow             2.2.0
tensorflow-estimator   2.2.0
termcolor              1.1.0
threadpoolctl          2.1.0
tifffile               2020.9.3
tk                     0.1.0
tr                     1.5.0
Twisted                18.9.0
ubuntu-advantage-tools 20.3
ufw                    0.36
unattended-upgrades    0.1
urllib3                1.25.8
wadllib                1.3.3
Werkzeug               1.0.1
wheel                  0.34.2
wrapt                  1.12.1
zipp                   1.0.0
zope.interface         4.7.1